+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/extradata/paradata/ParaDataUE.java
@@ -29,7 +29,7 @@ public class ParaDataUE {
 	@Getter
 	private List<Orchestrator> orchestrators = new ArrayList<>();
 	@Getter
-	private long collectionDateTimeStamp;
+	private long surveyValidationDateTimeStamp;
 
 	public List<ParadataVariable> getParadataVariable(String variableName) {
 		return this.paraDataVariables.get(variableName);
@@ -212,14 +212,14 @@ public class ParaDataUE {
 	}
 
 	/**
-	 * Gets the validation timestamp from events and put it into the collectionDateTimeStamp
+	 * Gets the validation timestamp from events and put it into the surveyValidationDateTimeStamp
 	 * @param validationParadataObjectId idParadataObject considered as validation event
 	 */
-	public void setCollectionDateTimeStamp(String validationParadataObjectId) {
+	public void setSurveyValidationDateTimeStamp(String validationParadataObjectId) {
 		long maxTimestamp = 0;
 		for(Event event : this.events){
 			if(event.getIdParadataObject().equals(validationParadataObjectId) && event.getTimestamp() >= maxTimestamp) {
-				this.collectionDateTimeStamp = event.getTimestamp();
+				this.surveyValidationDateTimeStamp = event.getTimestamp();
 				maxTimestamp = event.getTimestamp();
 			}
 		}